// apps/web/src/lib/api/generated/main/model/loots-controller-get-loot-stats-params.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */
import type { LootsControllerGetLootStatsPeriod } from './loots-controller-get-loot-stats-period';

export type LootsControllerGetLootStatsParams = {
period?: LootsControllerGetLootStatsPeriod;
world?: string;
npcTypes?: string;
excludeColossus?: boolean;
};